Added subscheck and test_subscheck
