+++ sympy/solvers/ode/__init__.py
@@ -1,6 +1,6 @@
-from .ode import (allhints, checkinfsol, checkodesol, classify_ode,
+from .ode import (allhints, checkinfsol, classify_ode,
         constantsimp, dsolve, homogeneous_order, infinitesimals)
-
+from .subscheck import checkodesol
 __all__ = [
     'allhints', 'checkinfsol', 'checkodesol', 'classify_ode', 'constantsimp',
     'dsolve', 'homogeneous_order', 'infinitesimals',
